server/requirements.txt
blinker==1.9.0
click==8.1.8
flask==3.1.0
flask-sock==0.7.0
future==1.0.0
h11==0.14.0
importlib-metadata==8.6.1
iso8601==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
pyserial==3.5
PyYAML==6.0.2
simple-websocket==1.1.0
six==1.17.0
sock==0.6.0
websocket-client==0.57.0
werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0